librlist: localize hwloc dependency in one file
Problem: it's hard to use anything in librlist without taking
on a libhwloc linkage requirement.

Move rlist_from_hwloc() to rhwloc.c.

Increase the scope of rlist_add_rnode() which was local to rlist.c
but is required by rlist_from_hwloc().  Define its prototype in
a new local header, rlist_private.h.

Move rhwloc.c from librlist.la to librlist-hwloc.la.
Explicitly link with librlist-hwloc.la where required.

Note: Since rlist_from_hwloc() is dropped from rlist.h, a python
ffi binding is no longer built for it.
