+++ src/common/librlist/rhwloc.c
@@ -23,7 +23,11 @@
 #include "src/common/libutil/read_all.h"
 #include "src/common/libutil/errno_safe.h"
 
+#include "rnode.h"
+#include "rlist.h"
+#include "rlist_private.h"
 #include "rhwloc.h"
+
 
 /*  Common hwloc_topology_init() and flags for Flux hwloc usage:
  */
@@ -296,6 +300,49 @@ char * rhwloc_gpu_idset_string (hwloc_topology_t topo)
         result = idset_encode (ids, IDSET_FLAG_RANGE);
     idset_destroy (ids);
     return result;
+}
+
+struct rlist *rlist_from_hwloc (int rank, const char *xml)
+{
+    char *ids = NULL;
+    struct rnode *n = NULL;
+    hwloc_topology_t topo = NULL;
+    const char *name;
+    struct rlist *rl = rlist_create ();
+
+    if (!rl)
+        return NULL;
+
+    if (xml)
+        topo = rhwloc_xml_topology_load (xml, RHWLOC_NO_RESTRICT);
+    else
+        topo = rhwloc_local_topology_load (0);
+    if (!topo)
+        goto fail;
+    if (!(ids = rhwloc_core_idset_string (topo))
+        || !(name = rhwloc_hostname (topo)))
+        goto fail;
+
+    if (!(n = rnode_create (name, rank, ids))
+        || rlist_add_rnode (rl, n) < 0)
+        goto fail;
+
+    free (ids);
+
+    if ((ids = rhwloc_gpu_idset_string (topo))
+        && rnode_add_child (n, "gpu", ids) < 0)
+        goto fail;
+
+    hwloc_topology_destroy (topo);
+    free (ids);
+    return rl;
+fail:
+    rlist_destroy (rl);
+    rnode_destroy (n);
+    free (ids);
+    if (topo)
+        hwloc_topology_destroy (topo);
+    return NULL;
 }
 
 /* vi: ts=4 sw=4 expandtab
